+++ frontend/src/App.jsx
@@ -1,11 +1,12 @@
 import { Route, Routes } from "react-router";
 import LoginPage from "./pages/LoginPage";
+import RecipeListPlaceholderPage from "./pages/RecipeListPlaceholderPage";
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<LoginPage />} />
-      {/* <Route path="/recipes" element={ } /> */}
+      <Route path="/recipes" element={<RecipeListPlaceholderPage />} />
     </Routes>
   )
 }